feat: show UAE introductory offer
